requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-CORS==3.0.10
Flask-Caching==1.10.1
Flask-WTF==0.14.3
python-dotenv==0.19.0
gunicorn==20.1.0
